refactor: utils.js をモジュール分割して src/utils/ に再配置
