+++ src/TrueTypeCollection.js
@@ -2,7 +2,7 @@ import * as r from 'restructure';
 import TTFFont from './TTFFont';
 import Directory from './tables/directory';
 import tables from './tables';
-import { asciiDecoder } from './utils';
+import { asciiDecoder } from './utils/decode';
 
 let TTCHeader = new r.VersionedStruct(r.uint32, {
   0x00010000: {
